backend/requirements.txt
Flask==3.0.2
azure-ai-formrecognizer==3.3.2
pillow==10.2.0
waitress==3.00